requirements.txt
django-cors-headers==3.13.0
djangorestframework==3.14.0

django-filter==22.1